core/database/main/requirements.txt
urllib3==1.26.11
networkx==3.0
requests==2.28.1
wget==3.2
colorama==0.4.5
click==8.1.3
graphqlclient==0.2.4
json-lines==0.5.0
websocket-client==1.7.0
pandas==2.2.2